// src/app/components/models/orders.model.ts
export interface OrdersModel{
  OrderNumber: string;
  OrderDueDate: string;
  CustomerBuyerName: string;
  CustomerAddress: string;
  CustomerPhone: string;
  OrderTotal: string;
}

// const ELEMENT_DATA: OrdersModel[] = [
//   {OrderNumber: 1, OrderDueDate: '19 Nov', CustomerBuyerName: 'sreekanth', CustomerAddress: '5-685 Hyderabad', CustomerPhone : 9876543210, OrderTotal : 4343},
//   {OrderNumber: 2, OrderDueDate: '34 dec', CustomerBuyerName: 'abc', CustomerAddress: '2-5 Delhi', CustomerPhone : 9876543210, OrderTotal : 4343},
//   {OrderNumber: 3, OrderDueDate: '30 june', CustomerBuyerName: 'def', CustomerAddress: '7-85 Hyderabad', CustomerPhone : 9876543210, OrderTotal : 4343},
//   {OrderNumber: 4, OrderDueDate: '9 may', CustomerBuyerName: 'ghi', CustomerAddress: '567-2 Hyderabad', CustomerPhone : 9876543210, OrderTotal : 4343},
//   {OrderNumber: 5, OrderDueDate: '30 feb', CustomerBuyerName: 'jkl', CustomerAddress: '849-4 Hyderabad', CustomerPhone : 9876543210, OrderTotal : 4343},
//   {OrderNumber: 6, OrderDueDate: '5 dec', CustomerBuyerName: 'mno', CustomerAddress: '78-0 Hyderabad', CustomerPhone : 9876543210, OrderTotal : 4343},
//   {OrderNumber: 7, OrderDueDate: '12 march', CustomerBuyerName: 'pqr', CustomerAddress: '948-5 Hyderabad', CustomerPhone : 9876543210, OrderTotal : 4343},
//   {OrderNumber: 8, OrderDueDate: '6Nov', CustomerBuyerName: 'stuvwxy', CustomerAddress: '5-685 Hyderabad', CustomerPhone : 9876543210, OrderTotal : 4343},
  
  
// ];